feat: Add configuration option to set taskbar as a win32 app bar
A new bar config option has been added which allows the user to specify if the bar should be treated as an Application Desktop Toolbar aka 'App Bar' by Windows.
Enabling this option will send a AppBarMessage to Windows, notifying the operating system to create a task bar area corresponding to the bar's geometry.
This means that the bar area is no-longer a part of the screen work area, and windwos can no longer be moved/dragged/maximised above the bar. Woohoo!